+++ NEWS.md
@@ -9,3 +9,4 @@
 - Decision Analysis plot size in PDF was too small
 - Replace histogram by distribution in regression plots
 - Fix bug in regex for overlapping labels in variables section of Model > Decision Analysis
+- Fixes for model with only an intercept (e.g., after stepwise regression)